make the entropy of the Universe a little bit smaller

+++ src/Menu/game_mode_selector.cpp
@@ -21,10 +21,10 @@ GameModeSelector::GameModeSelector(QWidget* parent, GameMode* game_mode) :
 }
 
 void GameModeSelector::PrepareComboBoxes() {
-  for (int i = 1; i <= kMaxPlayersAmount; i++) {
+  for (size_t i = 1; i <= kMaxPlayersAmount; i++) {
     number_of_players_->addItem(QString::number(i));
   }
-  for (int i = 1; i <= kMaxBotsAmount; i++) {
+  for (size_t i = 1; i <= kMaxBotsAmount; i++) {
     number_of_bots_->addItem(QString::number(i));
   }
 }
